+++ Runtime/ARFoundation/Textures/FixedLengthSwapchainStrategy.cs
@@ -16,8 +16,18 @@ namespace UnityEngine.XR.ARFoundation
     /// </remarks>
     class FixedLengthSwapchainStrategy : ISwapchainStrategy
     {
-        ARTextureInfo[][] m_TextureInfos;
-        readonly Dictionary<IntPtr, int> m_TextureInfoIndicesByNativePtr = new();
+        /// <summary>
+        /// A 2-dimensional array of texture info objects.
+        /// Each row represents a frame. Each column represents a texture that is used that frame.
+        /// </summary>
+        IUpdatableTexture[][] m_UpdatableTextures;
+
+        /// <summary>
+        /// Enables a constant-time lookup operation. Given the native pointer of any texture in the swapchain,
+        /// look up the index of the frame where that texture is used.
+        /// </summary>
+        readonly Dictionary<IntPtr, int> m_FrameIndicesByTexturePtr = new();
+
         readonly int m_NumTexturesPerFrame;
 
         internal FixedLengthSwapchainStrategy(NativeArray<NativeArray<XRTextureDescriptor>> swapchainDescriptors)
@@ -26,40 +36,41 @@ namespace UnityEngine.XR.ARFoundation
                 || swapchainDescriptors[0].Length == 0)
                 throw new ArgumentException("Swapchain must contain at least one texture", nameof(swapchainDescriptors));
 
-            int swapchainSize = swapchainDescriptors.Length;
+            int numFramesInSwapchain = swapchainDescriptors.Length;
             m_NumTexturesPerFrame = swapchainDescriptors[0].Length;
-            m_TextureInfos = new ARTextureInfo[swapchainSize][];
+            m_UpdatableTextures = new IUpdatableTexture[numFramesInSwapchain][];
 
-            for (var i = 0; i < swapchainSize; ++i)
+            for (var i = 0; i < numFramesInSwapchain; ++i)
             {
                 if (swapchainDescriptors[i].Length != m_NumTexturesPerFrame)
                     throw new ArgumentException(
                         "Swapchain must use the same number of textures per frame", nameof(swapchainDescriptors));
 
-                m_TextureInfos[i] = new ARTextureInfo[m_NumTexturesPerFrame];
+                m_UpdatableTextures[i] = new IUpdatableTexture[m_NumTexturesPerFrame];
                 for (var j = 0; j < m_NumTexturesPerFrame; ++j)
                 {
                     if (swapchainDescriptors[i][j].nativeTexture == IntPtr.Zero)
                         throw new ArgumentException("Swapchain textures must have non-null native texture pointer.", nameof(swapchainDescriptors));
 
-                    // Create all render textures if necessary
-                    m_TextureInfos[i][j] = new ARTextureInfo(swapchainDescriptors[i][j]);
+                    if (m_FrameIndicesByTexturePtr.ContainsKey(swapchainDescriptors[i][j].nativeTexture))
+                        throw new ArgumentException("Swapchain native texture pointers must be unique.", nameof(swapchainDescriptors));
 
-                    m_TextureInfoIndicesByNativePtr.Add(swapchainDescriptors[i][j].nativeTexture, i);
+                    m_UpdatableTextures[i][j] = UpdatableTextureFactory.Create(swapchainDescriptors[i][j]);
+                    m_FrameIndicesByTexturePtr.Add(swapchainDescriptors[i][j].nativeTexture, i);
                 }
             }
         }
 
-        bool ISwapchainStrategy.TryUpdateTextureInfosForFrame(
-            NativeArray<XRTextureDescriptor> descriptors, out ReadOnlyListSpan<ARTextureInfo> textureInfos)
+        bool ISwapchainStrategy.TryUpdateTexturesForFrame(
+            NativeArray<XRTextureDescriptor> descriptors, out ReadOnlyListSpan<IUpdatableTexture> textureInfos)
         {
             if (descriptors.Length == 0)
             {
-                textureInfos = ReadOnlyListSpan<ARTextureInfo>.Empty();
+                textureInfos = ReadOnlyListSpan<IUpdatableTexture>.Empty();
                 return true;
             }
 
-            if (!m_TextureInfoIndicesByNativePtr.TryGetValue(descriptors[0].nativeTexture, out var frameIndex))
+            if (!m_FrameIndicesByTexturePtr.TryGetValue(descriptors[0].nativeTexture, out var frameIndex))
                 throw new InvalidOperationException(
                     "Fixed-length swapchain strategy was given a texture that is not in the swapchain.");
 
@@ -70,19 +81,19 @@ namespace UnityEngine.XR.ARFoundation
             bool allRequiredTexturesDoExist = true;
             for (var j = 0; j < m_NumTexturesPerFrame; j++)
             {
-                if (!m_TextureInfos[frameIndex][j].TryUpdateTextureInfo(descriptors[j]))
+                if (!m_UpdatableTextures[frameIndex][j].TryUpdateFromDescriptor(descriptors[j]))
                     allRequiredTexturesDoExist = false;
             }
 
-            textureInfos = new ReadOnlyListSpan<ARTextureInfo>(m_TextureInfos[frameIndex]);
+            textureInfos = new ReadOnlyListSpan<IUpdatableTexture>(m_UpdatableTextures[frameIndex]);
             return allRequiredTexturesDoExist;
         }
 
         public void DestroyTextures()
         {
-            foreach (var frameArray in m_TextureInfos)
+            foreach (var frame in m_UpdatableTextures)
             {
-                foreach (var textureInfo in frameArray)
+                foreach (var textureInfo in frame)
                 {
                     textureInfo.DestroyTexture();
                 }